Add files via upload
working arduino sketch for arduino nano in my BMW 02-Series (E10) Dash
it monitors the control voltage going to the delay-off relay timer (HELLA 5HE 996.152-13). if the control voltage drops (ie turned off), the nano will initiate a short press (ie soft shutdown) or a long press (ie forced shutdown) depending if the Windows OS running simhub has already booted or not. This is to ensure a graceful shutdown whenever the power is turned off without first shutting down the OS.
The nano is connected via USB to Windows and a OS_Ready flag is set upon login of the user in Windows (autologon configured for simhub) via a python script called by Task Scheduler. The python script sets the OS_Ready flag on the arduino nano. If the flag isnt set, its assumed the OS hasnt booted yet.
Based on the flag, the arduino runs a short or long press. This is done via an external mosfet and 5V relay.