Add files via upload
python script for Windows to set the OS_Ready flag on the arduino nano connected via USB. this can be scheduled to run via task scheduler for the simhub autologin user for example. This ensures the USB has been initialised and connection has been established before the script is run. In my case, the nano is connected on COM5.